requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Image processing and computer vision
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-image>=0.18.0
imageio>=2.9.0

# Machine learning
scikit-learn>=1.0.0

# GPU acceleration
cupy-cuda11x>=9.0.0  # For CUDA 11.x, adjust based on your CUDA version
# cupy-cuda12x>=12.0.0  # Uncomment for CUDA 12.x

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Progress bars and utilities
tqdm>=4.62.0
colorama>=0.4.4

# File handling
zipfile36>=0.1.3
rarfile>=4.0

# Texture analysis
mahotas>=1.4.0
pyradiomics>=3.0.0

# Wavelets
PyWavelets>=1.1.0

# Additional utilities
psutil>=5.8.0
